fix: narrowing conversion warnings

+++ core/common/literals.hpp
@@ -13,23 +13,23 @@ namespace kagome::common::literals {
   }
 
   constexpr size_t operator""_kB(long double kilobytes) {
-    return (1ull << 10) * kilobytes;
+    return static_cast<size_t>((1ull << 10) * kilobytes);
   }
 
   constexpr size_t operator""_MB(long long unsigned int megabytes) {
-    return megabytes << 20u;
+    return static_cast<size_t>(megabytes << 20u);
   }
 
   constexpr size_t operator""_MB(long double megabytes) {
-    return (1ull << 20) * megabytes;
+    return static_cast<size_t>((1ull << 20) * megabytes);
   }
 
   constexpr size_t operator""_GB(long long unsigned int gigabytes) {
-    return gigabytes << 30u;
+    return static_cast<size_t>(gigabytes << 30u);
   }
 
   constexpr size_t operator""_GB(long double gigabytes) {
-    return (1ull << 30) * gigabytes;
+    return static_cast<size_t>((1ull << 30) * gigabytes);
   }
 
 }  // namespace kagome::common::literals
